Rename old template card to legacy
