+++ src/cards/legacy-template-card/const.ts
@@ -1,0 +1,4 @@
+import { PREFIX_NAME } from "../../const";
+
+export const LEGACY_TEMPLATE_CARD_NAME = `${PREFIX_NAME}-legacy-template-card`;
+export const LEGACY_TEMPLATE_CARD_EDITOR_NAME = `${LEGACY_TEMPLATE_CARD_NAME}-editor`;
