+++ src/cards/template-card/const.ts
@@ -1,4 +1,4 @@
 import { PREFIX_NAME } from "../../const";
 
 export const TEMPLATE_CARD_NAME = `${PREFIX_NAME}-template-card`;
-export const TEMPLATE_CARD_EDITOR_NAME = `${TEMPLATE_CARD_NAME}-editor`;
+export const TEMPLATE_CARD_EDITOR_NAME = `${TEMPLATE_CARD_NAME}-template-editor`;
