+++ src/cards/template-card/template-card.ts
@@ -1,43 +1,35 @@
 import { UnsubscribeFunc } from "home-assistant-js-websocket";
-import {
-  css,
-  CSSResultGroup,
-  html,
-  nothing,
-  PropertyValues,
-  TemplateResult,
-} from "lit";
+import { css, html, LitElement, nothing, PropertyValues } from "lit";
 import { customElement, property, state } from "lit/decorators.js";
 import { classMap } from "lit/directives/class-map.js";
+import { ifDefined } from "lit/directives/if-defined.js";
 import { styleMap } from "lit/directives/style-map.js";
+import memoizeOne from "memoize-one";
 import hash from "object-hash/dist/object_hash";
 import {
   actionHandler,
   ActionHandlerEvent,
-  computeRTL,
   handleAction,
   hasAction,
   HomeAssistant,
   LovelaceCard,
   LovelaceCardEditor,
   LovelaceGridOptions,
-  LovelaceLayoutOptions,
   RenderTemplateResult,
   subscribeRenderTemplate,
 } from "../../ha";
-import "../../shared/shape-icon";
-import "../../shared/state-info";
-import "../../shared/state-item";
-import { computeAppearance } from "../../utils/appearance";
-import { MushroomBaseElement } from "../../utils/base-element";
+import { computeCssColor } from "../../ha/common/color/compute-color";
+import { isTemplate } from "../../ha/common/string/has-template";
 import { CacheManager } from "../../utils/cache-manager";
-import { cardStyle } from "../../utils/card-styles";
-import { computeRgbColor } from "../../utils/colors";
 import { registerCustomCard } from "../../utils/custom-cards";
-import { getWeatherSvgIcon } from "../../utils/icons/weather-icon";
-import { weatherSVGStyles } from "../../utils/weather";
 import { TEMPLATE_CARD_EDITOR_NAME, TEMPLATE_CARD_NAME } from "./const";
 import { TemplateCardConfig } from "./template-card-config";
+
+registerCustomCard({
+  type: TEMPLATE_CARD_NAME,
+  name: "Mushroom Template",
+  description: "Build your own Mushroom card using templates",
+});
 
 const templateCache = new CacheManager<TemplateResults>(1000);
 
@@ -45,25 +37,26 @@ type TemplateResults = Partial<
   Record<TemplateKey, RenderTemplateResult | undefined>
 >;
 
-registerCustomCard({
-  type: TEMPLATE_CARD_NAME,
-  name: "Mushroom Template",
-  description: "Build your own mushroom card using templates",
-});
-
 const TEMPLATE_KEYS = [
   "icon",
-  "icon_color",
-  "badge_color",
-  "badge_icon",
+  "color",
   "primary",
   "secondary",
   "picture",
+  "badge_icon",
+  "badge_color",
+  "badge_text",
 ] as const;
+
 type TemplateKey = (typeof TEMPLATE_KEYS)[number];
 
+export interface LovelaceCardFeatureContext {
+  entity_id?: string;
+  area_id?: string;
+}
+
 @customElement(TEMPLATE_CARD_NAME)
-export class TemplateCard extends MushroomBaseElement implements LovelaceCard {
+export class Template extends LitElement implements LovelaceCard {
   public static async getConfigElement(): Promise<LovelaceCardEditor> {
     await import("./template-card-editor");
     return document.createElement(
@@ -71,16 +64,13 @@ export class TemplateCard extends MushroomBaseElement implements LovelaceCard {
     ) as LovelaceCardEditor;
   }
 
-  public static async getStubConfig(
-    _hass: HomeAssistant
-  ): Promise<TemplateCardConfig> {
+  public static getStubConfig(): TemplateCardConfig {
     return {
       type: `custom:${TEMPLATE_CARD_NAME}`,
-      primary: "Hello, {{user}}",
-      secondary: "How are you?",
-      icon: "mdi:home",
     };
   }
+
+  @property({ attribute: false }) public hass?: HomeAssistant;
 
   @state() private _config?: TemplateCardConfig;
 
@@ -90,79 +80,6 @@ export class TemplateCard extends MushroomBaseElement implements LovelaceCard {
     TemplateKey,
     Promise<UnsubscribeFunc>
   > = new Map();
-
-  @property({ reflect: true, type: String })
-  public layout: string | undefined;
-
-  public getCardSize(): number | Promise<number> {
-    let height = 1;
-    if (!this._config) return height;
-    const appearance = computeAppearance(this._config);
-    if (appearance.layout === "vertical") {
-      height += 1;
-    }
-    return height;
-  }
-
-  public getLayoutOptions(): LovelaceLayoutOptions {
-    const options: LovelaceLayoutOptions = {
-      grid_columns: 2,
-      grid_rows: 1,
-    };
-    if (!this._config) return options;
-    const appearance = computeAppearance(this._config);
-    if (appearance.layout === "vertical") {
-      options.grid_rows! += 1;
-    }
-    if (appearance.layout === "horizontal") {
-      options.grid_columns = 4;
-    }
-    if (this._config?.multiline_secondary) {
-      options.grid_rows = undefined;
-    }
-    return options;
-  }
-
-  // For HA < 2024.11
-  public getGridOptions(): LovelaceGridOptions {
-    // No min and max because the content can be dynamic
-    const options: LovelaceGridOptions = {
-      columns: 6,
-      rows: 1,
-    };
-    if (!this._config) return options;
-    const appearance = computeAppearance(this._config);
-    if (appearance.layout === "vertical") {
-      options.rows! += 1;
-    }
-    if (appearance.layout === "horizontal") {
-      options.columns = 12;
-    }
-    if (this._config?.multiline_secondary) {
-      options.rows = undefined;
-    }
-    return options;
-  }
-
-  setConfig(config: TemplateCardConfig): void {
-    TEMPLATE_KEYS.forEach((key) => {
-      if (
-        this._config?.[key] !== config[key] ||
-        this._config?.entity != config.entity
-      ) {
-        this._tryDisconnectKey(key);
-      }
-    });
-    this._config = {
-      tap_action: {
-        action: "toggle",
-      },
-      hold_action: {
-        action: "more-info",
-      },
-      ...config,
-    };
-  }
 
   public connectedCallback() {
     super.connectedCallback();
@@ -199,126 +116,6 @@ export class TemplateCard extends MushroomBaseElement implements LovelaceCard {
     }
   }
 
-  private _handleAction(ev: ActionHandlerEvent) {
-    handleAction(this, this.hass!, this._config!, ev.detail.action!);
-  }
-
-  public isTemplate(key: TemplateKey) {
-    const value = this._config?.[key];
-    return value?.includes("{");
-  }
-
-  private getValue(key: TemplateKey) {
-    return this.isTemplate(key)
-      ? this._templateResults?.[key]?.result?.toString()
-      : this._config?.[key];
-  }
-
-  protected render() {
-    if (!this._config || !this.hass) {
-      return nothing;
-    }
-
-    const icon = this.getValue("icon");
-    const iconColor = this.getValue("icon_color");
-    const badgeIcon = this.getValue("badge_icon");
-    const badgeColor = this.getValue("badge_color");
-    const primary = this.getValue("primary");
-    const secondary = this.getValue("secondary");
-    const picture = this.getValue("picture");
-
-    const multiline_secondary = this._config.multiline_secondary;
-
-    const rtl = computeRTL(this.hass);
-
-    const appearance = computeAppearance({
-      fill_container: this._config.fill_container,
-      layout: this._config.layout,
-      icon_type: Boolean(picture)
-        ? "entity-picture"
-        : Boolean(icon)
-          ? "icon"
-          : "none",
-      primary_info: Boolean(primary) ? "name" : "none",
-      secondary_info: Boolean(secondary) ? "state" : "none",
-    });
-
-    const weatherSvg = getWeatherSvgIcon(icon);
-
-    return html`
-      <ha-card
-        class=${classMap({ "fill-container": appearance.fill_container })}
-      >
-        <mushroom-card .appearance=${appearance} ?rtl=${rtl}>
-          <mushroom-state-item
-            ?rtl=${rtl}
-            .appearance=${appearance}
-            @action=${this._handleAction}
-            .actionHandler=${actionHandler({
-              hasHold: hasAction(this._config.hold_action),
-              hasDoubleClick: hasAction(this._config.double_tap_action),
-            })}
-          >
-            ${picture
-              ? this.renderPicture(picture)
-              : weatherSvg
-                ? html`<div slot="icon">${weatherSvg}</div>`
-                : icon
-                  ? this.renderIcon(icon, iconColor)
-                  : nothing}
-            ${(icon || picture) && badgeIcon
-              ? this.renderBadgeIcon(badgeIcon, badgeColor)
-              : undefined}
-            <mushroom-state-info
-              slot="info"
-              .primary=${primary}
-              .secondary=${secondary}
-              .multiline_secondary=${multiline_secondary}
-            ></mushroom-state-info>
-          </mushroom-state-item>
-        </mushroom-card>
-      </ha-card>
-    `;
-  }
-
-  renderPicture(picture: string): TemplateResult {
-    return html`
-      <mushroom-shape-avatar
-        slot="icon"
-        .picture_url=${(this.hass as any).hassUrl(picture)}
-      ></mushroom-shape-avatar>
-    `;
-  }
-
-  renderIcon(icon: string, iconColor?: string) {
-    const iconStyle = {};
-    if (iconColor) {
-      const iconRgbColor = computeRgbColor(iconColor);
-      iconStyle["--icon-color"] = `rgb(${iconRgbColor})`;
-      iconStyle["--shape-color"] = `rgba(${iconRgbColor}, 0.2)`;
-    }
-    return html`
-      <mushroom-shape-icon style=${styleMap(iconStyle)} slot="icon">
-        <ha-state-icon .hass=${this.hass} .icon=${icon}></ha-state-icon>
-      </mushroom-shape-icon>
-    `;
-  }
-
-  renderBadgeIcon(badge: string, badgeColor?: string) {
-    const badgeStyle = {};
-    if (badgeColor) {
-      const iconRgbColor = computeRgbColor(badgeColor);
-      badgeStyle["--main-color"] = `rgba(${iconRgbColor})`;
-    }
-    return html`
-      <mushroom-badge-icon
-        slot="badge"
-        .icon=${badge}
-        style=${styleMap(badgeStyle)}
-      ></mushroom-badge-icon>
-    `;
-  }
-
   protected updated(changedProps: PropertyValues): void {
     super.updated(changedProps);
     if (!this._config || !this.hass) {
@@ -326,6 +123,13 @@ export class TemplateCard extends MushroomBaseElement implements LovelaceCard {
     }
 
     this._tryConnect();
+  }
+
+  private _getTemplateKeyValue(key: TemplateKey): string {
+    if (!this._config) {
+      return "";
+    }
+    return this._config[key] ?? "";
   }
 
   private async _tryConnect(): Promise<void> {
@@ -338,12 +142,15 @@ export class TemplateCard extends MushroomBaseElement implements LovelaceCard {
     if (
       this._unsubRenderTemplates.get(key) !== undefined ||
       !this.hass ||
-      !this._config ||
-      !this.isTemplate(key)
+      !this._config
     ) {
       return;
     }
 
+    const value = this._getTemplateKeyValue(key);
+    if (!isTemplate(value)) {
+      return;
+    }
     try {
       const sub = subscribeRenderTemplate(
         this.hass.connection,
@@ -354,12 +161,13 @@ export class TemplateCard extends MushroomBaseElement implements LovelaceCard {
           };
         },
         {
-          template: this._config[key] ?? "",
+          template: value,
           entity_ids: this._config.entity_id,
           variables: {
             config: this._config,
             user: this.hass.user!.name,
             entity: this._config.entity,
+            area: this._config.area,
           },
           strict: true,
         }
@@ -408,25 +216,361 @@ export class TemplateCard extends MushroomBaseElement implements LovelaceCard {
     }
   }
 
-  static get styles(): CSSResultGroup {
-    return [
-      super.styles,
-      cardStyle,
-      css`
-        mushroom-state-item {
-          cursor: pointer;
-        }
-        mushroom-shape-icon {
-          --icon-color: rgb(var(--rgb-disabled));
-          --shape-color: rgba(var(--rgb-disabled), 0.2);
-        }
-        svg {
-          width: var(--icon-size);
-          height: var(--icon-size);
-          display: flex;
-        }
-        ${weatherSVGStyles}
-      `,
-    ];
+  public setConfig(config: TemplateCardConfig): void {
+    this._config = {
+      ...config,
+    };
+    if (this._config.icon_color) {
+      this._config.color = this._config.icon_color;
+      delete this._config.icon_color;
+    }
   }
+
+  private _featureContext = memoizeOne(
+    (config: TemplateCardConfig): LovelaceCardFeatureContext => {
+      return {
+        entity_id: config.entity,
+        area_id: config.area,
+      };
+    }
+  );
+
+  private getValue(key: TemplateKey) {
+    const value = this._getTemplateKeyValue(key);
+    return isTemplate(value)
+      ? this._templateResults?.[key]?.result?.toString()
+      : value;
+  }
+
+  public getCardSize(): number {
+    const featuresPosition =
+      this._config && this._featurePosition(this._config);
+    const featuresCount = this._config?.features?.length || 0;
+    return (
+      1 +
+      (this._config?.vertical ? 1 : 0) +
+      (featuresPosition === "inline" ? 0 : featuresCount)
+    );
+  }
+
+  public getGridOptions(): LovelaceGridOptions {
+    const columns = 6;
+    let min_columns = 6;
+    let rows: number | undefined = 1;
+    const featurePosition = this._config && this._featurePosition(this._config);
+    const featuresCount = this._config?.features?.length || 0;
+    if (featuresCount) {
+      if (featurePosition === "inline") {
+        min_columns = 12;
+      } else {
+        rows += featuresCount;
+      }
+    }
+
+    if (this._config?.vertical) {
+      if (
+        this._config.primary ||
+        (this._config.secondary && !this._config.icon)
+      ) {
+        rows++;
+      }
+    }
+    if (this._config?.multiline_secondary) {
+      rows = undefined;
+    }
+    return {
+      columns,
+      rows,
+    };
+  }
+
+  private _handleAction(ev: ActionHandlerEvent) {
+    handleAction(this, this.hass!, this._config!, ev.detail.action!);
+  }
+
+  private _handleIconAction(ev: CustomEvent) {
+    ev.stopPropagation();
+    const config = {
+      entity: this._config!.entity,
+      tap_action: this._config!.icon_tap_action,
+      hold_action: this._config!.icon_hold_action,
+      double_tap_action: this._config!.icon_double_tap_action,
+    };
+    handleAction(this, this.hass!, config, ev.detail.action!);
+  }
+
+  private get _hasCardAction() {
+    return (
+      hasAction(this._config?.tap_action) ||
+      hasAction(this._config?.hold_action) ||
+      hasAction(this._config?.double_tap_action)
+    );
+  }
+
+  private get _hasIconAction() {
+    return (
+      hasAction(this._config?.icon_tap_action) ||
+      hasAction(this._config?.icon_hold_action) ||
+      hasAction(this._config?.icon_double_tap_action)
+    );
+  }
+
+  private _featurePosition = memoizeOne((config: TemplateCardConfig) => {
+    if (config.vertical) {
+      return "bottom";
+    }
+    return config.features_position || "bottom";
+  });
+
+  private _displayedFeatures = memoizeOne((config: TemplateCardConfig) => {
+    const features = config.features || [];
+    const featurePosition = this._featurePosition(config);
+
+    if (featurePosition === "inline") {
+      return features.slice(0, 1);
+    }
+    return features;
+  });
+
+  protected render() {
+    if (!this._config || !this.hass) {
+      return nothing;
+    }
+    const entityId = this._config.entity;
+
+    const contentClasses = { vertical: Boolean(this._config.vertical) };
+
+    const icon = this.getValue("icon");
+    const color = this.getValue("color");
+    const cssColor = color ? computeCssColor(color) : "";
+    const primary = this.getValue("primary");
+    const secondary = this.getValue("secondary");
+    const picture = this.getValue("picture");
+    const badgeIcon = this.getValue("badge_icon");
+    const badgeColor = this.getValue("badge_color");
+    const badgeText = this.getValue("badge_text");
+    const badgeCssColor = badgeColor ? computeCssColor(badgeColor) : "";
+
+    const style = {
+      "--tile-color": cssColor,
+    };
+
+    const featurePosition = this._featurePosition(this._config);
+    const features = this._displayedFeatures(this._config);
+
+    const containerOrientationClass =
+      featurePosition === "inline" ? "horizontal" : "";
+
+    const multilineSecondary = this._config.multiline_secondary;
+
+    const secondaryInfo = secondary
+      ? html`
+          <span
+            style=${styleMap({
+              "white-space": multilineSecondary ? "pre-wrap" : "nowrap",
+            })}
+            >${secondary}</span
+          >
+        `
+      : undefined;
+
+    const featureContext = this._featureContext(this._config);
+
+    return html`
+      <ha-card style=${styleMap(style)}>
+        <div
+          class="background"
+          @action=${this._handleAction}
+          .actionHandler=${actionHandler({
+            disabled: !this._hasCardAction,
+            hasHold: hasAction(this._config!.hold_action),
+            hasDoubleClick: hasAction(this._config!.double_tap_action),
+          })}
+          role=${ifDefined(this._hasCardAction ? "button" : undefined)}
+          tabindex=${ifDefined(this._hasCardAction ? "0" : undefined)}
+          aria-labelledby="info"
+        >
+          <ha-ripple .disabled=${!this._hasCardAction}></ha-ripple>
+        </div>
+        <div class="container ${containerOrientationClass}">
+          <div class="content ${classMap(contentClasses)}">
+            ${icon
+              ? html`
+                  <ha-tile-icon
+                    role=${ifDefined(
+                      this._hasIconAction ? "button" : undefined
+                    )}
+                    tabindex=${ifDefined(this._hasIconAction ? "0" : undefined)}
+                    @action=${this._handleIconAction}
+                    .actionHandler=${actionHandler({
+                      disabled: !this._hasIconAction,
+                      hasHold: hasAction(this._config!.icon_hold_action),
+                      hasDoubleClick: hasAction(
+                        this._config!.icon_double_tap_action
+                      ),
+                    })}
+                    .interactive=${this._hasIconAction}
+                    .imageUrl=${picture}
+                  >
+                    <ha-state-icon
+                      slot="icon"
+                      .icon=${icon}
+                      .hass=${this.hass}
+                    ></ha-state-icon>
+                    ${badgeIcon || badgeText
+                      ? html`
+                          <ha-tile-badge
+                            style=${styleMap({
+                              "--badge-color": badgeCssColor,
+                            })}
+                          >
+                            ${badgeText
+                              ? html`
+                                  <span style="line-height: 1;">
+                                    ${badgeText}
+                                  </span>
+                                `
+                              : html`<ha-icon .icon=${badgeIcon}> </ha-icon>`}
+                          </ha-tile-badge>
+                        `
+                      : nothing}
+                  </ha-tile-icon>
+                `
+              : nothing}
+            ${primary || secondary
+              ? html`
+                  <ha-tile-info
+                    id="info"
+                    .primary=${primary}
+                    .secondary=${secondaryInfo}
+                  ></ha-tile-info>
+                `
+              : nothing}
+          </div>
+          ${features.length > 0
+            ? html`
+                <hui-card-features
+                  .hass=${this.hass}
+                  .context=${featureContext}
+                  .color=${cssColor}
+                  .features=${features}
+                  .position=${featurePosition}
+                ></hui-card-features>
+              `
+            : nothing}
+        </div>
+      </ha-card>
+    `;
+  }
+
+  static styles = css`
+    :host {
+      --tile-color: var(--state-inactive-color);
+      -webkit-tap-highlight-color: transparent;
+    }
+    ha-card:has(.background:focus-visible) {
+      --shadow-default: var(--ha-card-box-shadow, 0 0 0 0 transparent);
+      --shadow-focus: 0 0 0 1px var(--tile-color);
+      border-color: var(--tile-color);
+      box-shadow: var(--shadow-default), var(--shadow-focus);
+    }
+    ha-card {
+      --ha-ripple-color: var(--tile-color);
+      --ha-ripple-hover-opacity: 0.04;
+      --ha-ripple-pressed-opacity: 0.12;
+      height: 100%;
+      transition:
+        box-shadow 180ms ease-in-out,
+        border-color 180ms ease-in-out;
+      display: flex;
+      flex-direction: column;
+      justify-content: space-between;
+    }
+    [role="button"] {
+      cursor: pointer;
+      pointer-events: auto;
+    }
+    [role="button"]:focus {
+      outline: none;
+    }
+    .background {
+      position: absolute;
+      top: 0;
+      left: 0;
+      bottom: 0;
+      right: 0;
+      border-radius: var(--ha-card-border-radius, 12px);
+      margin: calc(-1 * var(--ha-card-border-width, 1px));
+      overflow: hidden;
+    }
+    .container {
+      margin: calc(-1 * var(--ha-card-border-width, 1px));
+      display: flex;
+      flex-direction: column;
+      flex: 1;
+    }
+    .container.horizontal {
+      flex-direction: row;
+    }
+
+    .content {
+      position: relative;
+      display: flex;
+      flex-direction: row;
+      align-items: center;
+      padding: 10px;
+      flex: 1;
+      min-width: 0;
+      box-sizing: border-box;
+      pointer-events: none;
+      gap: 10px;
+    }
+
+    .vertical {
+      flex-direction: column;
+      text-align: center;
+      justify-content: center;
+    }
+    .vertical ha-tile-info {
+      width: 100%;
+      flex: none;
+    }
+    ha-tile-icon {
+      --tile-icon-color: var(--tile-color);
+      position: relative;
+      padding: 6px;
+      margin: -6px;
+    }
+    ha-tile-badge {
+      position: absolute;
+      top: 3px;
+      right: 3px;
+      inset-inline-end: 3px;
+      inset-inline-start: initial;
+      --tile-badge-background-color: var(
+        --badge-color,
+        var(--secondary-text-color)
+      );
+    }
+    ha-tile-info {
+      position: relative;
+      min-width: 0;
+      transition: background-color 180ms ease-in-out;
+      box-sizing: border-box;
+    }
+    hui-card-features {
+      --feature-color: var(--tile-color);
+      padding: 0 12px 12px 12px;
+    }
+    .container.horizontal hui-card-features {
+      width: calc(50% - var(--column-gap, 0px) / 2 - 12px);
+      flex: none;
+      --feature-height: 36px;
+      padding: 0 12px;
+      padding-inline-start: 0;
+    }
+    .secondary.multiline {
+      white-space: pre-wrap;
+    }
+  `;
 }
