--- src/cards/template-new-card/const.ts
@@ -1,4 +1,0 @@
-import { PREFIX_NAME } from "../../const";
-
-export const TEMPLATE_CARD_NEW_NAME = `${PREFIX_NAME}-template-new-card`;
-export const TEMPLATE_CARD_NEW_EDITOR_NAME = `${TEMPLATE_CARD_NEW_NAME}-template-new-editor`;
