+++ src/mushroom.ts
@@ -13,6 +13,7 @@ import "./cards/empty-card/empty-card";
 import "./cards/entity-card/entity-card";
 import "./cards/fan-card/fan-card";
 import "./cards/humidifier-card/humidifier-card";
+import "./cards/legacy-template-card/legacy-template-card";
 import "./cards/light-card/light-card";
 import "./cards/lock-card/lock-card";
 import "./cards/media-player-card/media-player-card";
@@ -20,7 +21,6 @@ import "./cards/number-card/number-card";
 import "./cards/person-card/person-card";
 import "./cards/select-card/select-card";
 import "./cards/template-card/template-card";
-import "./cards/template-new-card/template-new-card";
 import "./cards/title-card/title-card";
 import "./cards/update-card/update-card";
 import "./cards/vacuum-card/vacuum-card";
